requirements.txt
asgiref==3.3.4
Django==3.2.3
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
